djangosite-master/requirements.txt
asgiref==3.2.10
astroid==2.4.2
bootstrap4==0.1.0
boto==2.49.0
boto3==1.16.16
botocore==1.19.16
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-dotenv==1.4.2
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-restframework==0.0.1
django-storages==1.10.1
environ==1.0
gunicorn==20.0.4
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
package-name==0.1
phonenumbers==8.12.14
phonenumberslite==8.12.15
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.1
python-dotenv==0.15.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1